data block creator code cleanup
